+++ src/blackbench/micro-targets/nested.py
@@ -5,6 +5,13 @@ def func(
     lots_of_paren = (((3,((3,(((1,( 2,(31,((4,(5,(4),5)))),4,))),54,),)), 4),),)
 
 
+def func(
+    arg1: Tuple[Tuple[List[str], bool], Sequence[Set[str]]],
+    arg2: generate_class(generate_name(get_user(get_system())))
+) -> Union[Literal["no"], str, bool, generate_class(get_weather())]:
+    lots_of_paren = (((3,((3,(((1,( 2,(31,((4,(5,(4),5)))),4,))),54,),)), 4),),)
+
+
 def level_one():
     def level_two():
         def level_three():
@@ -30,6 +37,39 @@ def level_one():
     ...
 
 
+def level_one():
+    def level_two():
+        def level_three():
+            def level_four():
+                def level_five():
+                    def level_six():
+                        def level_seven():
+                            def level_eight():
+                                def level_nine():
+                                    def level_ten():
+                                        def level_eleven():
+                                            if (num * ((a**l)- (a + ((((((b))))))+(e)) + 9)) != eleven:
+                                                turn_it_up_to_eleven()
+                                        ...
+                                    ...
+                                ...
+                            ...
+                        ...
+                    ...
+                ...
+            ...
+        ...
+    ...
+
+
+
+if oss == "free":
+    if oss == "open":
+        if oss == "distributed":
+            if oss == "democratic":
+                print("oss doesn't mean corps just get to expect EoL'd Pythons to work for free")
+                print(f"{f'{{{{{{{hi_fellow_maintainers}}}}}}}'}")
+
 if oss == "free":
     if oss == "open":
         if oss == "distributed":
@@ -39,3 +79,16 @@ if oss == "free":
 
 
 my_list = [*[*[*[*[*[*[*[133], 123],1927],410],100], 371], 401], 163]
+my_list = [*[*[*[*[*[*[*[133], 123],1927],410],100], 371], 401], 163]
+my_list = [*[*[*[*[*[*[*[133], 123],1927],410],100], 371], 401], 163]
+
+
+i_am_pretty_sure_this_is_nested_at_the_ast_level = (
+    1 / 23 / 3/ 4 / 232 / 342342342 / 3423 / 234 / 3 / 2423 / 5 / 23/ 4 / 234 / 234 / 23 / 42323 / 423 / 423 / 4 / 23 / 42
+)
+i_am_pretty_sure_this_is_nested_at_the_ast_level = (
+    1 / 23 / 3/ 4 / 232 / 342342342 / 3423 / 234 / 3 / 2423 / 5 / 23/ 4 / 234 / 234 / 23 / 42323 / 423 / 423 / 4 / 23 / 42
+)
+i_am_pretty_sure_this_is_nested_at_the_ast_level = (
+    1 / 23 / 3/ 4 / 232 / 342342342 / 3423 / 234 / 3 / 2423 / 5 / 23/ 4 / 234 / 234 / 23 / 42323 / 423 / 423 / 4 / 23 / 42
+)