+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 21.7a1
+## Unreleased
 
 Date of release: *n/a*